[update] データ関連をひとまとめ
イベント関連のクラスをDataフォルダへ移動、スキルをEventLibraryクラスに移動
エラー時のメッセージ変更、エラー時は起動しないように
