+++ UmaUmaChecker/UmaUmaChecker.vcxproj
@@ -244,11 +244,11 @@
     <ClCompile Include="src\UI\CheckUpdateDialog.cpp" />
     <ClCompile Include="src\Recognizer\CombineImage.cpp" />
     <ClCompile Include="src\UI\DebugImageCombineFrame.cpp" />
-    <ClCompile Include="src\Loader\Event\EventData.cpp" />
+    <ClCompile Include="src\Data\EventData.cpp" />
     <ClCompile Include="src\UI\FontComboBox.cpp" />
     <ClCompile Include="src\UI\GrandLiveMusicListFrame.cpp" />
     <ClCompile Include="src\Log\Log.cpp" />
-    <ClCompile Include="src\Loader\Event\ScenarioData.cpp" />
+    <ClCompile Include="src\Data\ScenarioData.cpp" />
     <ClCompile Include="src\Recognizer\ScrollbarDetector.cpp" />
     <ClCompile Include="src\Tesseract\Tesseract.cpp" />
     <ClCompile Include="src\UI\Theme\ThemeWrapper.cpp" />
@@ -266,9 +266,9 @@
     <ClCompile Include="src\UI\PreviewFrame.cpp" />
     <ClCompile Include="src\UI\SettingDialog.cpp" />
     <ClCompile Include="src\System\SystemUsage.cpp" />
-    <ClCompile Include="src\Loader\Event\TextCollector.cpp" />
+    <ClCompile Include="src\Data\TextCollector.cpp" />
     <ClCompile Include="src\Recognizer\Uma.cpp" />
-    <ClCompile Include="src\Loader\Event\EventLibrary.cpp" />
+    <ClCompile Include="src\Data\EventLibrary.cpp" />
     <ClCompile Include="src\UI\MainFrame.cpp" />
     <ClCompile Include="src\Utils\utility.cpp" />
     <ClCompile Include="src\UI\wxTextPopupCtrl.cpp" />
@@ -288,13 +288,13 @@
     <ClInclude Include="src\UI\CheckUpdateDialog.h" />
     <ClInclude Include="src\Recognizer\CombineImage.h" />
     <ClInclude Include="src\UI\DebugImageCombineFrame.h" />
-    <ClInclude Include="src\Loader\Event\EventData.h" />
+    <ClInclude Include="src\Data\EventData.h" />
     <ClInclude Include="src\UI\FontComboBox.h" />
     <ClInclude Include="src\UI\GrandLiveMusicListFrame.h" />
     <ClInclude Include="src\Log\Log.h" />
     <ClInclude Include="src\object_pool.hpp" />
     <ClInclude Include="src\Point.h" />
-    <ClInclude Include="src\Loader\Event\ScenarioData.h" />
+    <ClInclude Include="src\Data\ScenarioData.h" />
     <ClInclude Include="src\Recognizer\ScrollbarDetector.h" />
     <ClInclude Include="src\Tesseract\Tesseract.h" />
     <ClInclude Include="src\UI\Theme\ThemeWrapper.h" />
@@ -310,12 +310,12 @@
     <ClInclude Include="src\UI\DebugFrame.h" />
     <ClInclude Include="src\UI\PreviewFrame.h" />
     <ClInclude Include="resource.h" />
-    <ClInclude Include="src\Loader\Event\Event.h" />
+    <ClInclude Include="src\Data\Event.h" />
     <ClInclude Include="src\UI\SettingDialog.h" />
     <ClInclude Include="src\System\SystemUsage.h" />
-    <ClInclude Include="src\Loader\Event\TextCollector.h" />
+    <ClInclude Include="src\Data\TextCollector.h" />
     <ClInclude Include="src\Recognizer\Uma.h" />
-    <ClInclude Include="src\Loader\Event\EventLibrary.h" />
+    <ClInclude Include="src\Data\EventLibrary.h" />
     <ClInclude Include="src\UI\MainFrame.h" />
     <ClInclude Include="src\Utils\utility.h" />
     <ClInclude Include="src\version.h" />
