+++ UmaUmaChecker/src/Data/TextCollector.h
